moved functions to Datalink.ts
